Use props in screen.js
